+++ packages/stack-server/sentry.edge.config.ts
@@ -13,4 +13,6 @@ Sentry.init({
 
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
+
+  enabled: process.env.NODE_ENV !== "development" && !process.env.CI,
 });