+++ packages/stack-server/sentry.server.config.ts
@@ -13,7 +13,5 @@ Sentry.init({
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
 
-  // uncomment the line below to enable Spotlight (https://spotlightjs.com)
-  // spotlight: process.env.NODE_ENV === 'development',
-  
+  enabled: process.env.NODE_ENV !== "development" && !process.env.CI,
 });