+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # stack-docs
 
+## 0.0.5
+
+### Patch Changes
+
+- CommonJS support
+
 ## 0.0.4
 
 ### Patch Changes