+++ packages/stack-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack-server
 
+## 0.3.7
+
+### Patch Changes
+
+- CommonJS support
+- Updated dependencies
+  - @stackframe/stack-shared@2.3.5
+  - @stackframe/stack@2.3.7
+
 ## 0.3.6
 
 ### Patch Changes